refactor: ♻️ Refactor SeedData and register new migration AddContactInfo

+++ sources/JobMagnet/Infrastructure/Entities/ProfileEntity.cs
@@ -26,7 +26,7 @@ public class ContactInfoEntity : SoftDeletableEntity<long>
 {
     public required string Value { get; set; }
 
-    [ForeignKey(nameof(ContactType))] public long ContactTypeId { get; set; }
+    [ForeignKey(nameof(ContactType))] public int ContactTypeId { get; set; }
     [ForeignKey(nameof(Resume))] public long ResumeId { get; set; }
 
     public virtual ContactTypeEntity ContactType { get; set; }
